+++ package.json
@@ -13,8 +13,11 @@
   },
   "dependencies": {
     "@heroicons/react": "^2.1.1",
+    "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react": "^18.2.0"
+    "react-router": "^6.22.0",
+    "react-router-config": "^5.1.1",
+    "react-router-dom": "^6.22.0"
   },
   "devDependencies": {
     "@types/react": "^18.2.55",
