+++ src/main.tsx
@@ -1,38 +1,20 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./components/App.tsx";
-import { useRegisterSW } from "virtual:pwa-register/react";
+import { RouterProvider, createBrowserRouter } from 'react-router-dom'
+import App from "./routes/App.tsx";
+import AppUpdater from "./components/AppUpdater.tsx";
 import "./index.css";
 
-const AppUpdater = () => {
-  const intervalMS = 1000 * 60 * 60 // 1 hour;
-
-  useRegisterSW({
-    onRegisteredSW(swUrl, r) {
-      r &&
-        setInterval(async () => {
-          if (!(!r.installing && navigator)) return;
-          if ("connection" in navigator && !navigator.onLine) return;
-
-          const resp = await fetch(swUrl, {
-            cache: "no-store",
-            headers: {
-              cache: "no-store",
-              "cache-control": "no-cache",
-            },
-          });
-
-          if (resp?.status === 200) await r.update();
-        }, intervalMS);
-    },
-  });
-
-  return null;
-};
+const router = createBrowserRouter([
+  {
+    path: "/",
+    element: <App />
+  },
+]);
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
-    <App />
+    <RouterProvider router={router} />
     <AppUpdater />
   </React.StrictMode>,
 );
